Move easydb files to its own folders
